+++ subgraph-js/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This CHANGELOG pertains only to Apollo Federation packages in the 2.x range. The Federation v0.x equivalent for this package can be found [here](https://github.com/apollographql/federation/blob/version-0.x/subgraph-js/CHANGELOG.md) on the `version-0.x` branch of this repo.
 
+## vNext
+
+- Fix output of `printSubgraphSchema` method, ensuring it can be read back by composition and `buildSubgraphSchema` [PR #1831](https://github.com/apollographql/federation/pull/1831).
+- Fix definition of `@key` to be repeatable [PR #1826](https://github.com/apollographql/federation/pull/1826).
+
 ## 2.0.2-alpha.1
 
 - Add __resolveType to _Entity union [PR #1773](https://github.com/apollographql/federation/pull/1773)